+++ p2p/metrics.gen.go
@@ -68,6 +68,12 @@ func PrometheusMetrics(namespace string, labelsAndValues ...string) *Metrics {
 			Name:      "messages_reactor_in_flight",
 			Help:      "Number of messages in flight (wip by reactor)",
 		}, append(labels, "message_type", "reactor")).With(labelsAndValues...),
+		MessagesReactorDropped: prometheus.NewCounterFrom(stdprometheus.CounterOpts{
+			Namespace: namespace,
+			Subsystem: MetricsSubsystem,
+			Name:      "messages_reactor_dropped",
+			Help:      "Number of messages dropped because the reactor's priority queue was full",
+		}, append(labels, "message_type", "reactor")).With(labelsAndValues...),
 		MessagesReactorPendingDuration: prometheus.NewHistogramFrom(stdprometheus.HistogramOpts{
 			Namespace: namespace,
 			Subsystem: MetricsSubsystem,
@@ -100,6 +106,7 @@ func NopMetrics() *Metrics {
 		MessageSendBytesTotal:          discard.NewCounter(),
 		MessagesReceived:               discard.NewCounter(),
 		MessagesReactorInFlight:        discard.NewGauge(),
+		MessagesReactorDropped:         discard.NewCounter(),
 		MessagesReactorPendingDuration: discard.NewHistogram(),
 		MessageReactorReceiveDuration:  discard.NewHistogram(),
 		MessageReactorQueueConcurrency: discard.NewGauge(),